IDEA Linux Push

+++ src/main/java/maxdistructo/droidbot2/background/Listener.java
@@ -21,7 +21,7 @@ import static maxdistructo.droidbot2.BaseBot.client;
 public class Listener {
     public static boolean blackJackRunning = false;
     public static String blackjackAnswer;
-    public static String prefix = "d!"; //Change back to ! for release!
+    public static String prefix = "!"; //Change back to ! for release!
     public static String triviaAnswer;
 
     @EventSubscriber
@@ -46,46 +46,47 @@ public class Listener {
         }
         String content = message.getContent();
         Object messageContent[] = content.split(" ");
-       // if (Perms.checkOwner(message)) { //REMOVE THIS FOR RELEASE!!
-          //  if (messageContent[0].equals(prefix + "bj")) { //broken
-          //      message.reply(BlackJack.onBlackjackCommand(messageContent, message));
-          //  }
+
+        if(message.getGuild().getLongID() == L) {
+            //  if (messageContent[0].equals(prefix + "bj")) { //broken
+            //      message.reply(BlackJack.onBlackjackCommand(messageContent, message));
+            //  }
             //if(messageContent[0].equals("hit") || messageContent[0].equals("stay") || messageContent[0].equals("double") && blackJackRunning){
-           //     blackjackAnswer = (String)messageContent[0];
-          //  }
+            //     blackjackAnswer = (String)messageContent[0];
+            //  }
             if (messageContent[0].equals(prefix + "check")) { //Works
-                Message.sendMessage(message.getChannel(), Message.simpleEmbed(message.getAuthor(), "Check", Check.onCheckCommand(messageContent, message),message));
+                Message.sendMessage(message.getChannel(), Message.simpleEmbed(message.getAuthor(), "Check", Check.onCheckCommand(messageContent, message), message));
                 message.delete();
             } else if (messageContent[0].equals(prefix + "casino")) { //Works except for admin commands
-                message.reply("",Message.simpleEmbed(message.getAuthor(), "Casino",Casino.onCasinoCommand(messageContent,message,message.getAuthor()),message));
+                message.reply("", Message.simpleEmbed(message.getAuthor(), "Casino", Casino.onCasinoCommand(messageContent, message, message.getAuthor()), message));
                 message.delete();
             } else if (messageContent[0].equals(prefix + "50") || messageContent[0].equals(prefix + "fifty")) { //Works
-                message.reply("",Message.simpleEmbed(message.getAuthor(), "FiftyFifty", FiftyFifty.onFiftyCommand(messageContent, message), message));
+                message.reply("", Message.simpleEmbed(message.getAuthor(), "FiftyFifty", FiftyFifty.onFiftyCommand(messageContent, message), message));
                 message.delete();
             } else if (messageContent[0].equals(prefix + "fortune")) { //Works
-                message.reply("",Message.simpleEmbed(message.getAuthor(),"Fortune",Fortune.onFortuneCommand(messageContent, message),message));
+                message.reply("", Message.simpleEmbed(message.getAuthor(), "Fortune", Fortune.onFortuneCommand(messageContent, message), message));
                 message.delete();
             }
-                //else if(messageContent[0].equals(prefix + "game")){ //Broke
-                //  message.reply(GameCommand.onGameCommand(messageContent, message));
-           // }
+            //else if(messageContent[0].equals(prefix + "game")){ //Broke
+            //  message.reply(GameCommand.onGameCommand(messageContent, message));
+            // }
             else if (messageContent[0].equals(prefix + "info")) { //Not Working ATM
-                Message.sendMessage(message.getChannel(), Message.simpleEmbed(message.getAuthor(),"Info", Info.onInfoCommand(messageContent, message, mentioned),message));
+                Message.sendMessage(message.getChannel(), Message.simpleEmbed(message.getAuthor(), "Info", Info.onInfoCommand(messageContent, message, mentioned), message));
                 message.delete();
             } else if (messageContent[0].equals(prefix + "insult")) { //Works
-                Message.sendMessage(message.getChannel(), Message.simpleEmbed(message.getAuthor(),"Insult", Insult.onInsultCommand(messageContent, message, mentioned), message));
+                Message.sendMessage(message.getChannel(), Message.simpleEmbed(message.getAuthor(), "Insult", Insult.onInsultCommand(messageContent, message, mentioned), message));
                 message.delete();
             } else if (messageContent[0].equals(prefix + "debug")) { //Needs perms set.
-                message.reply("",Message.simpleEmbed(message.getAuthor(),"Debug", Debug.onDebugCommand((String[]) messageContent, message),message));
+                message.reply("", Message.simpleEmbed(message.getAuthor(), "Debug", Debug.onDebugCommand((String[]) messageContent, message), message));
                 message.delete();
             } else if (messageContent[0].equals(prefix + "shutdown")) { //Works
-                message.reply("",Message.simpleEmbed(message.getAuthor(),"Shutdown",Shutdown.onShutdownCommand(message),message));
+                message.reply("", Message.simpleEmbed(message.getAuthor(), "Shutdown", Shutdown.onShutdownCommand(message), message));
                 message.delete();
             } else if (messageContent[0].equals(prefix + "help")) {
                 Message.sendDM(message.getAuthor(), Help.onHelpCommand());
                 message.delete();
             } else if (messageContent[0].equals(prefix + "allin")) {
-                message.reply("",Message.simpleEmbed(message.getAuthor(),"Allin",Allin.onAllinCommand(messageContent, message),message));
+                message.reply("", Message.simpleEmbed(message.getAuthor(), "Allin", Allin.onAllinCommand(messageContent, message), message));
                 message.delete();
             } else if (messageContent[0].equals(prefix + "say") && channelMention != null) {
                 Message.sendMessage(channelMention, Say.onSayCommand(messageContent, message, channelMention));
@@ -94,40 +95,30 @@ public class Listener {
             } else if (messageContent[0].equals(prefix + "spam")) {
                 Message.sendMessage(message.getChannel(), Spam.onSpamCommand(messageContent, message, mentioned));
                 message.delete();
-            }
-            else if (messageContent[0].equals(prefix + "slap")){
-                Message.sendMessage(message.getChannel(), PlayerFun.onSlapCommand(message,mentioned));
-            }
-            else if (messageContent[0].equals(prefix + "tnt")){
-                Message.sendMessage(message.getChannel(), PlayerFun.onTntCommand(message,mentioned));
-            }
-            else if (messageContent[0].equals(prefix + "kiss")){
-                Message.sendMessage(message.getChannel(), PlayerFun.onKissCommand(message,mentioned));
-            }
-            else if (messageContent[0].equals(prefix + "hug")){
-                Message.sendMessage(message.getChannel(), PlayerFun.onHugCommand(message,mentioned));
-            }
-            else if (messageContent[0].equals(prefix + "poke")){
-                Message.sendMessage(message.getChannel(), PlayerFun.onPokeCommand(message,mentioned));
-            }
-            else if (messageContent[0].equals(prefix + "respect") || messageContent[0].equals("/f")){
-                Message.sendMessage(message.getChannel(), PlayerFun.onPayRespects(message,mentioned));
+            } else if (messageContent[0].equals(prefix + "slap")) {
+                Message.sendMessage(message.getChannel(), PlayerFun.onSlapCommand(message, mentioned));
+            } else if (messageContent[0].equals(prefix + "tnt")) {
+                Message.sendMessage(message.getChannel(), PlayerFun.onTntCommand(message, mentioned));
+            } else if (messageContent[0].equals(prefix + "kiss")) {
+                Message.sendMessage(message.getChannel(), PlayerFun.onKissCommand(message, mentioned));
+            } else if (messageContent[0].equals(prefix + "hug")) {
+                Message.sendMessage(message.getChannel(), PlayerFun.onHugCommand(message, mentioned));
+            } else if (messageContent[0].equals(prefix + "poke")) {
+                Message.sendMessage(message.getChannel(), PlayerFun.onPokeCommand(message, mentioned));
+            } else if (messageContent[0].equals(prefix + "respect") || messageContent[0].equals("/f")) {
+                Message.sendMessage(message.getChannel(), PlayerFun.onPayRespects(message, mentioned));
                 message.delete();
+            } else if (messageContent[0].equals(prefix + "banhammer")) {
+                Message.sendMessage(message.getChannel(), PlayerFun.onBanHammer(message, mentioned));
             }
-            else if(messageContent[0].equals(prefix + "banhammer")){
-                Message.sendMessage(message.getChannel(), PlayerFun.onBanHammer(message,mentioned));
-            }
-          //   else if(messageContent[0].equals(prefix + "trivia")) {
-         //       message.reply(Trivia.onTriviaCommand(messageContent, message));
-          //  }
-          //  else if(messageContent[0].equals(triviaAnswer)){
-          //      Trivia.addTriviaScore(event.getMessage());
-          //      Trivia.checkTrivia(event.getMessage());
-         //   }
-
-      //  } else if (message.getContent().contains("d@")) {
-      //      message.reply("Command Error: User has attempted to run a WIP command.");
-      //  }
+            //   else if(messageContent[0].equals(prefix + "trivia")) {
+            //       message.reply(Trivia.onTriviaCommand(messageContent, message));
+            //  }
+            //  else if(messageContent[0].equals(triviaAnswer)){
+            //      Trivia.addTriviaScore(event.getMessage());
+            //      Trivia.checkTrivia(event.getMessage());
+            //   }
+        }
     }
